requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
python-dotenv>=1.0.0
pydantic>=2.10.0
detoxify>=0.5.2
torch>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
